Remove IDisposeAble from recovery
